rename model to types
